requirements.txt
nfcpy
requests
soco